requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.26.3
scikit-learn>=1.4.0
matplotlib>=3.8.2
seaborn>=0.13.1
pydantic>=2.5.3
python-multipart>=0.0.6
fastapi>=0.109.0
uvicorn>=0.27.0
